Updated project outline and removing trash files
I updated my part of the project outline and made some minor changes
on Juan's part. We can always revert it if we want. We just need to
read it together to make sure everybody is happy with it.

I delete some unnecessary files from the github tree. We don't need
to track files generated by the LaTeX compiler or temp Matlab files.
(I'm speaking to you, Juan)

Bellow is a list of the changed files:

	deleted:    outlineCompartingMethods.tex
	deleted:    project-outline/project_outline.aux
	deleted:    project-outline/project_outline.log
	modified:   project-outline/project_outline.pdf
	modified:   project-outline/project_outline.tex
	deleted:    project_outline.pdf
	deleted:    project_outline.tex